Remove the patch pinning Ansible version to 6.0 in CNS
There was an issue in cloud-native-stack that required
Ansible 6 instead of the newly released Ansible 7. That issue
has been fixed in cloud-native-stack so the patch is being
removed.

Additionally, an "apt update" has been added before the installation
of cloud-native-stack to make sure that standard packages like
python3-pip are available. This fixes some spurious failures
that were seen from the Ubuntu package manager.